+++ mininterface/textual_interface/textual_facet.py
@@ -1,4 +1,5 @@
 from typing import TYPE_CHECKING
+from warnings import warn
 from ..facet import Facet
 if TYPE_CHECKING:
     from .textual_adaptor import TextualAdaptor
@@ -15,8 +16,16 @@ class TextualFacet(Facet):
     # NOTE: multiline title will not show up
     def set_title(self, title: str):
         self._title = title
-        self.adaptor.app.title = title
+        try:
+            self.adaptor.app.title = title
+        except:
+            # NOTE: When you receive Facet in Command.init, the app does not exist yet
+            warn("Setting textual title not implemented well.")
 
     def submit(self, *args, **kwargs):
         super().submit(*args, **kwargs)
-        self.adaptor.app.action_confirm()
+        try:
+            self.adaptor.app.action_confirm()
+        except:
+            # NOTE: When you receive Facet in Command.init, the app does not exist yet
+            warn("Setting textual title not implemented well.")